+++ app/components/wallet/ada-redemption/AdaRedemptionChoices.js
@@ -20,6 +20,16 @@ const messages = defineMessages({
     defaultMessage: '!!!Paper vended',
     description: 'Tab title "Paper vended" on Ada redemption page.'
   },
+  recoveryRegularTabTitle: {
+    id: 'wallet.redeem.choices.tab.title.recoveryRegular',
+    defaultMessage: '!!!Recovery - regular',
+    description: 'Tab title "Recovery - regular" on Ada redemption page.'
+  },
+  recoveryForceVendedTabTitle: {
+    id: 'wallet.redeem.choices.tab.title.recoveryForceVended',
+    defaultMessage: '!!!Recovery - force vended',
+    description: 'Tab title "Recovery - force vended" on Ada redemption page.'
+  },
 });
 
 type Props = {
@@ -56,6 +66,18 @@ export default class AdaRedemptionChoices extends Component<Props> {
           onClick={() => onSelectChoice('paperVended')}
         >
           {intl.formatMessage(messages.paperVendedTabTitle)}
+        </button>
+        <button
+          className={activeChoice === 'recoveryRegular' ? styles.activeButton : ''}
+          onClick={() => onSelectChoice('recoveryRegular')}
+        >
+          {intl.formatMessage(messages.recoveryRegularTabTitle)}
+        </button>
+        <button
+          className={activeChoice === 'recoveryForceVended' ? styles.activeButton : ''}
+          onClick={() => onSelectChoice('recoveryForceVended')}
+        >
+          {intl.formatMessage(messages.recoveryForceVendedTabTitle)}
         </button>
       </div>
     );