+++ app/stores/ada/AdaRedemptionStore.js
@@ -30,6 +30,7 @@ export default class AdaRedemptionStore extends Store {
   @observable email: ?string = null;
   @observable adaPasscode: ?string = null;
   @observable adaAmount: ?string = null;
+  @observable decryptionKey: ?string = null;
   @observable redemptionCode: string = '';
   @observable walletId: ?string = null;
   @observable error: ?LocalizableError = null;
@@ -49,6 +50,7 @@ export default class AdaRedemptionStore extends Store {
     actions.setEmail.listen(this._setEmail);
     actions.setAdaPasscode.listen(this._setAdaPasscode);
     actions.setAdaAmount.listen(this._setAdaAmount);
+    actions.setDecryptionKey.listen(this._setDecryptionKey);
     actions.redeemAda.listen(this._redeemAda);
     actions.redeemPaperVendedAda.listen(this._redeemPaperVendedAda);
     actions.adaSuccessfullyRedeemed.listen(this._onAdaSuccessfullyRedeemed);
@@ -128,6 +130,11 @@ export default class AdaRedemptionStore extends Store {
     this._parseCodeFromCertificate();
   });
 
+  _setDecryptionKey = action(({ decryptionKey } : { decryptionKey: string }) => {
+    this.decryptionKey = decryptionKey;
+    this._parseCodeFromCertificate();
+  });
+
   _parseCodeFromCertificate() {
     if (this.redemptionType === 'regular') {
       if (!this.passPhrase && this.isCertificateEncrypted) return;
@@ -137,16 +144,28 @@ export default class AdaRedemptionStore extends Store {
         return;
       }
     }
+    if (this.redemptionType === 'recoveryRegular') {
+      if (!this.decryptionKey && this.isCertificateEncrypted) return;
+    }
+    if (this.redemptionType === 'recoveryForceVended') {
+      if (!this.passPhrase && this.isCertificateEncrypted) return;
+    }
     if (this.redemptionType === 'paperVended') return;
     if (this.certificate == null) throw new Error('Certificate File is required for parsing.');
     const path = this.certificate.path; // eslint-disable-line
     Logger.debug('Parsing ADA Redemption code from certificate: ' + path);
     let decryptionKey = null;
-    if (this.redemptionType === 'regular' && this.isCertificateEncrypted) {
+    if (
+      (this.redemptionType === 'regular' || this.redemptionType === 'recoveryForceVended') &&
+      this.isCertificateEncrypted
+    ) {
       decryptionKey = this.passPhrase;
     }
     if (this.redemptionType === 'forceVended' && this.isCertificateEncrypted) {
       decryptionKey = [this.email, this.adaPasscode, this.adaAmount];
+    }
+    if (this.redemptionType === 'recoveryRegular' && this.isCertificateEncrypted) {
+      decryptionKey = this.decryptionKey;
     }
     ipcRenderer.send(PARSE_REDEMPTION_CODE.REQUEST, path, decryptionKey, this.redemptionType);
   }